requirements.txt
altair==4.2.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
<<<<<<< HEAD
clip==1.0
colorama==0.4.6
decorator==5.1.1
entrypoints==0.4
faiss-cpu==1.7.3
filelock==3.10.4
=======
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cmake==3.26.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
faiss-cpu==1.7.3
filelock==3.10.3
fonttools==4.39.2
>>>>>>> a0fb0c3620c0013faad6f467799ff91504fb30e3
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.1.0
<<<<<<< HEAD
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
=======
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lit==16.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
>>>>>>> a0fb0c3620c0013faad6f467799ff91504fb30e3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
<<<<<<< HEAD
=======
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
>>>>>>> a0fb0c3620c0013faad6f467799ff91504fb30e3
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
<<<<<<< HEAD
=======
pyparsing==3.0.9
>>>>>>> a0fb0c3620c0013faad6f467799ff91504fb30e3
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
rich==13.3.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.20.0
sympy==1.11.1
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
<<<<<<< HEAD
typing_extensions==4.5.0
tzdata==2023.2
=======
triton==2.0.0
typing_extensions==4.5.0
tzdata==2022.7
>>>>>>> a0fb0c3620c0013faad6f467799ff91504fb30e3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.15.0
